+++ lib/motion_toast.dart
@@ -442,28 +442,36 @@ class MotionToast extends StatefulWidget {
   /// default `= true`
   final bool displaySideBar;
 
-  late BuildContext _context;
+  //Overlay that does not block the screen
+  OverlayEntry? overlayEntry;
 
   /// Display the created motion toast based on the [position] attribute
   /// [context]: the actual context of the application
   void show(BuildContext context) {
-    _context = context;
-    Navigator.of(context).push(
-      PageRouteBuilder<Widget>(
-        fullscreenDialog: false,
-        barrierColor: barrierColor,
-        pageBuilder: (BuildContext context, _, __) => this,
-        opaque: false,
-        barrierDismissible: dismissable,
-      ),
-    );
+    overlayEntry = _overlayEntryBuilder();
+    Overlay.maybeOf(context)?.insert(overlayEntry!);
   }
 
-  void dismiss() {
-    if (currentWidgetState.mounted) {
-      Navigator.of(_context).pop();
-      onClose?.call();
-    }
+  void closeOverlay() {
+    overlayEntry?.remove();
+    overlayEntry = null;
+  }
+
+  OverlayEntry _overlayEntryBuilder() {
+    return OverlayEntry(
+      opaque: false,
+      builder: (context) {
+        return SafeArea(
+          child: AlertDialog(
+            backgroundColor: Colors.transparent,
+            contentPadding: const EdgeInsets.all(0),
+            insetPadding: const EdgeInsets.all(30),
+            elevation: 0,
+            content: this,
+          ),
+        );
+      },
+    );
   }
 }
 
@@ -475,7 +483,7 @@ class _MotionToastState extends State<MotionToast>
 
   void _popCurrentToast() {
     if (mounted) {
-      Navigator.of(context).pop();
+      widget.closeOverlay();
       widget.onClose?.call();
     }
   }